+++ packages/client/verifiable-credentials/src/types/verifiableCredential.ts
@@ -18,13 +18,3 @@ export interface EncryptedVerifiableCredential {
 }
 
 export type VerifiableCredentialType = VerifiableCredential | EncryptedVerifiableCredential;
-
-export function isEncryptedVerifiableCredential(
-    credential: VerifiableCredentialType
-): credential is EncryptedVerifiableCredential {
-    return (
-        (credential as EncryptedVerifiableCredential).id !== undefined &&
-        (credential as EncryptedVerifiableCredential).payload !== undefined &&
-        (credential as VerifiableCredential).credentialSubject === undefined
-    );
-}
